--- tests/trainer/flags/test_logger.py
@@ -1,36 +1,0 @@
-# Copyright The PyTorch Lightning team.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-import pytest
-
-from pytorch_lightning import Trainer
-
-
-@pytest.mark.parametrize(
-    "logger_value",
-    [
-        None,
-        False,
-        [],
-    ],
-)
-def test_no_logger(tmpdir, logger_value):
-    """Test the cases where logger=None, logger=False, logger=[] are passed to Trainer."""
-    trainer = Trainer(
-        logger=logger_value,
-        default_root_dir=tmpdir,
-        max_steps=1,
-    )
-    assert trainer.logger is None
-    assert trainer.loggers == []
-    assert trainer.log_dir == tmpdir
